Disable sja1105_spi polling while configuring switch
The kernel polling the switch while it was being reconfigured caused the
switch configuration to sporadically fail. Use the poll disable
mechanism in sja1105_spi to prevent this.

For other SPI drivers this ioctl will fail as unsupported, but this can
be safely ignored as those drivers do no polling anyways.